backend/requirements.txt
asgiref==3.4.1
astroid==2.8.3
certifi==2021.5.30
charset-normalizer==2.0.6
coreapi==2.3.3
coreschema==0.0.4
Django==3.2.7
django-cors-headers==3.10.0
django-debug-toolbar==3.2.2
django-nested-inline==0.4.4
django-ordered-model==3.4.3
djangorestframework==3.12.4
drf-nested-routers==0.93.4
drf-yasg==1.20.0
feedparser==6.0.8
gunicorn==20.1.0
idna==3.2
inflection==0.5.1
isort==5.9.3
itypes==1.2.0
Jinja2==3.0.2
lazy-object-proxy==1.6.0
MarkupSafe==2.0.1
mccabe==0.6.1
packaging==21.0
Pillow==8.3.2
platformdirs==2.4.0
psycopg2==2.9.1
pylint==2.11.1
pyparsing==2.4.7
pysendpulse==0.1.2
python3-memcached==1.51
pytz==2021.3
requests==2.26.0
ruamel.yaml==0.17.16
ruamel.yaml.clib==0.2.6
sgmllib3k==1.0.0
six==1.16.0
sqlparse==0.4.2
toml==0.10.2
transliterate==1.10.2
typing-extensions==3.10.0.2
uritemplate==3.0.1
urllib3==1.26.7
uuid==1.30
wrapt==1.13.2
